// src/app/interfaces/player.ts
export interface Player {
    id?: string;
    name: string;
    hole1: number;
    hole2: number;
    hole3: number;
    hole4: number;
    hole5: number;
    hole6: number;
    hole7: number;
    hole8: number;
    hole9: number;
    hole10: number;
    hole11: number;
    hole12: number;
    hole13: number;
    hole14: number;
    hole15: number;
    hole16: number;
    hole17: number;
    hole18: number;
      
}